Moved IP Publisher and updated launch files
